fix a lot of commands, more error reporting. 7 cmd and 2 event to go

+++ bot/cogs/cleanup.py
@@ -5,7 +5,7 @@ CivvieBot cog to handle cleanup of stale games from the database.
 import logging
 from datetime import datetime
 from time import time
-from pony.orm import db_session
+from pony.orm import db_session, left_join
 from discord.ext import commands, tasks
 from database.models import Player, Game, WebhookURL
 from utils import config
@@ -24,16 +24,50 @@ class Cleanup(commands.Cog):
         '''
         self.bot: commands.Bot = bot
         # Apparently this isn't wrapped.
-        self.cleanup.start() # pylint: disable=no-member
+        self.run_cleanup.start() # pylint: disable=no-member
 
 
     @tasks.loop(seconds=config.get('cleanup_interval'))
-    async def cleanup(self):
+    async def run_cleanup(self):
         '''
         Cleans up stale games, delete-flagged URLs, and delete-flagged players.
         '''
-        cleanup_limit = config.get('cleanup_limit')
-        await self.remove_stale_games(cleanup_limit)
+        await self.cleanup(self.bot)
+    
+
+    @staticmethod
+    async def cleanup(bot: commands.Bot, limit_channel: int = None):
+        '''
+        Cleans up games, players, and webhook URLs.
+        '''
+        now = time()
+        limit = config.get('cleanup_limit')
+        stale_len = config.get('stale_game_length')
+        # Cheaper than hanging onto a ton of entities.
+        games_removed = 0
+        players_removed = 0
+        urls_removed = 0
+        with db_session():
+            games = (Game.select(lambda g: g.lastturn + stale_len < now 
+                and g.webhookurl.channelid == limit_channel)[:limit]
+                if limit_channel else
+                Game.select(lambda g: g.lastturn + stale_len < now)[:limit])
+            for game in games:
+                last_turn = datetime.fromtimestamp(
+                    game.lastturn).strftime('%m/%%d/%Y, %H:%M:%S')
+                game.delete()
+                games_removed += 1
+                logger.info(
+                    'Deleted game %s (%d) and associated players during cleanup (last turn: %s)',
+                    game.gamename,
+                    game.id,
+                    last_turn)
+                channel = await bot.fetch_channel(game.webhookurl.id)
+                await channel.send((f'No activity detected in the game {game.gamename} for '
+                    f'{expand_seconds_to_string(stale_len)} (last turn: {last_turn}), so'
+                    'information about the game has been removed. If you would like to continue '
+                    'recieving notifications for this game, a new turn will have to be taken and '
+                    'CivvieBot will have to recieve a turn notification for it'))
         # Pony appears to have inaccurate documentation regarding bulk delete;
         # we should be able to attach .delete(bulk=True) to the end of the
         # select, but db.EntityMeta.select() doesn't return the type of object
@@ -42,37 +76,21 @@ class Cleanup(commands.Cog):
         # troubling. If it continues, a rework using a non-ORM query builder
         # like PyPika may be worth considering.
         with db_session():
-            for player in Player.select(lambda p: p.cleanup == True)[:cleanup_limit]:
+            for player in (left_join(p for p in Player for g in p.games 
+                if p.cleanup == True
+                and g.webhookurl.channelid == limit_channel)[:limit]
+                if limit_channel else Player.select(lambda p: p.cleanup == True)[:limit]):
                 player.delete()
-            for url in WebhookURL.select(lambda w: w.cleanup == True)[:cleanup_limit]:
-                url.delete()
+                players_removed += 1
+            if not limit_channel:
+                for url in WebhookURL.select(lambda w: w.cleanup == True)[:limit]:
+                    url.delete()
+                    urls_removed += 1
 
-
-    async def remove_stale_games(self, limit: int):
-        '''
-        Removes up to limit stale games from the database.
-
-        A game is considered stale if the current time is greater than the sum of the game's last
-        turn and the stale_game_length set in configuration.
-        '''
-        now = time()
-        stale_len = config.get('stale_game_length')
-        with db_session():
-            for game in Game.select(lambda g: g.lastturn + stale_len < now)[:limit]:
-                last_turn = datetime.fromtimestamp(
-                    game.lastturn).strftime('%m/%%d/%Y, %H:%M:%S')
-                game.delete()
-                logger.info(
-                    'Deleted game %s (%d) and associated players during cleanup (last turn: %s)',
-                    game.gamename,
-                    game.id,
-                    last_turn)
-                channel = await self.bot.fetch_channel(game.webhookurl.id)
-                await channel.send((f'No activity detected in the game {game.gamename} for '
-                    f'{expand_seconds_to_string(stale_len)} (last turn: {last_turn}), so'
-                    'information about the game has been removed. If you would like to continue '
-                    'recieving notifications for this game, a new turn will have to be taken and '
-                    'CivvieBot will have to recieve a turn notification for it'))
+        logger.info('Round of cleanup has finished; removed %d games, %d players, %d URLs',
+            games_removed,
+            players_removed,
+            urls_removed)
 
 
 def setup(bot: commands.Bot):
